Put strings used by ccall metadata in AS 0.
This is needed because the metadata structures go in a comdat and so
will be deduplicated, then the strings that they refer to will be
dropped by the linker as unreferenced.  If they are in AS200, then each
version of the comdat  will *also* end up in the __cap_relocs section
and the linker will complain.